requirements.txt
absl-py==2.3.1
accelerate==1.8.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
bitarray==3.5.0
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
braceexpand==0.1.7
cachetools==6.1.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cloudpickle==3.1.1
colorama==0.4.6
confection==0.1.5
contexttimer==0.3.3
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
Cython==3.1.2
datasets==3.6.0
decorator==5.2.1
decord==0.6.0
deepspeed==0.17.1
diffusers==0.34.0
dill==0.3.8
distlib==0.3.9
distro==1.9.0
einops==0.8.1
evaluate==0.4.4
exceptiongroup==1.3.0
executing==2.2.0
fairscale==0.4.4
fairseq==0.12.2
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
hydra-core==1.0.7
identify==2.6.12
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
iopath==0.1.10
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
jsonlines==4.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kaggle==1.7.4.5
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.1
lightning-utilities==0.14.3
llvmlite==0.44.0
lxml==6.0.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
modelscope==1.28.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.13
narwhals==1.46.0
networkx==3.3
ninja==1.11.1.4
nltk==3.9.1
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.575.51
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.0.6
openai==1.93.3
opencv-python-headless==4.12.0.88
opendatasets==0.1.22
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
pooch==1.8.2
portalocker==3.2.0
pre_commit==4.2.0
preshed==3.0.10
prompt_toolkit==3.0.51
promptcap==1.0.3
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycocoevalcap==1.2
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyDeprecate==0.3.2
pydot==4.0.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-slugify==8.0.4
pytorch-lightning==1.7.7
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rouge-score==0.0.4
rpds-py==0.26.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
sacrebleu==2.5.1
safetensors==0.5.3
salesforce-lavis==1.0.0
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
sentencepiece==0.2.0
sentry-sdk==2.32.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart_open==7.3.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
streamlit==1.46.1
sympy==1.13.3
tabulate==0.9.0
taming-transformers-rom1504==0.0.6
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.16
tokenizers==0.21.2
toml==0.10.2
torchmetrics==1.7.4
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.0
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
unicodedata2==16.0.0
urllib3==2.5.0
virtualenv==20.31.2
wandb==0.21.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webdataset==1.0.2
webencodings==0.5.1
Werkzeug==3.1.3
word2number==1.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zhconv==1.4.3
zipp==3.23.0
